Remove default symbol evaluation
Referenced symbols should always refer to a bound variable in the scope
hierarchy. Subtle bugs could be introduced with the existing strategy.
Instead, you must `quote` a symbol to avoid evaluating it. This made
some of the tests functionally wrong, so they needed updating.

This also adds the `t` and `nil` sentinels, and printing empty lists as
`nil` instead. This communicates the use in both directions (you can
construct an empty list with `nil`, `(list)` or `()`).

`boundp` is also added to avoid tests that use the error machinery for
validation because I'm not happy with the errors as-is. I'll need to
peak at the common Lisp implementation for some inspiration, here.
Bubbling errors as values is not ideal, as the trace/source is not
clear.