Refactor lists into cons cells
This is technically a more correct model, even though it was a huge
refactor. There are some new footguns, like calling `expect_cons_cell`
before checking if the current value is `nil`. Might be best to have a
variant that allows nil to exist/be mapped.

Overall, the code has been reduced, but there is some space for
improvement on the native function validation.

I've also raised the error type into the same concept as a Rust error,
however this has dropped the concept of a stack trace and made it harder
to debug. It may be worth designing a version of the interpreter which
tracks the function call stack (by name, if any, ignoring fully
anonymous lambdas).